move all caches and functions into internal API
