+++ internal-api/src/main/java/datadog/trace/api/IntFunction.java
@@ -1,0 +1,6 @@
+package datadog.trace.api;
+
+// TODO replace me when baselining against JDK8
+public interface IntFunction<T> {
+  T apply(int value);
+}
